Refactor: Service seperation
